tests: Adding 1 and 1000 commits samples
To make the tests more easy to read, a file containing a simple commit is added
to test only the json importing.

'sample_commits' is renamed to 'sample_1000_commits' and adds much more lines.
This will be useful to get more commits for testing the code in a reproducible
way.