git_stats: Fixing authors sorting
The sorting code was broken as Counter() is not sorting the output by default.
It was pretty a chance the actual test output was sorted but failed on some
other.

This patch uses the most_common() properties of Counter to avoid this side
effect.

Adding a test case to avoid this again.